requirements.txt
PyMuPDF
langchain
faiss-cpu
openai
numpy
python-dotenv
langchain-community
tiktoken
pyyaml
pytest
black
isort
mypy
langchain_openai
transformers 
torch
torchvision 
torchaudio